+++ src/main/java/kr/codesqaud/cafe/board/dto/PostResponseForm.java
@@ -1,26 +1,21 @@
-package kr.codesqaud.cafe.board.domain;
-
-import kr.codesqaud.cafe.board.dto.PostResponseForm;
+package kr.codesqaud.cafe.board.dto;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class BoardPost {
+public class PostResponseForm {
     private Long postId;
     private String writer;
     private String title;
     private String contents;
     private LocalDateTime writeDateTime;
 
-    public BoardPost(String writer, String title, String contents, LocalDateTime writeDateTime) {
+    public PostResponseForm(Long postId, String writer, String title, String contents, LocalDateTime writeDateTime) {
+        this.postId = postId;
         this.writer = writer;
         this.title = title;
         this.contents = contents;
         this.writeDateTime = writeDateTime;
-    }
-
-    public PostResponseForm toPostResponseForm() {
-        return new PostResponseForm(postId, writer, title, contents, writeDateTime);
     }
 
     public Long getPostId() {
